Refactorin auth package : code simplification + typo correction (displayLoginView)
